@AlstonLin 2017-11-01
@macisamuele The performance changes occurred when the logging level was not in DEBUG, (i removed log.debug, so it should not of logged), although I'm sure there would be a very significant performance hit if I did set the logging level to that. I think the performance hit comes from calculating the params to format the string and actually formatting the string itself.