@macisamuele 2017-11-02
@AlstonLin thanks a lot for your contributions, it's really appreciated.

I like the idea of replacing the way ``deref`` operates by overriding deref completely since it's "easy" and provides an additional 3% performance improvements, but I have some notes about how to achieve this result.

You're silently changing ``deref`` method while executing ``_validate_config`` method, which seems really strange to me.
What if we're having ``deref`` method that calls an inner method that will be the original ``deref`` or the NO_OP deref? This should continue to give us performance improvements while ensuring future code readability (an example of this could be ``_internal_spec_dict``).

I have in mind something like
```python
def _deref(self, ref_dict):
    # Original spec.deref implementation

@cached_property
def _deref_method(self):
     if self.config['internally_dereference_refs']:
        return lambda ref_dict: ref_dict
    else:
        return self._deref

def deref(self, ref_dict):
    return self._deref_method()(ref_dict)
```

Extra note. Are you able to publish benchmark results and comparison of the ``master`` branch and your're branch? It will keep trace and (publicly available) data about the additional performance improvements.
Hint. Running ``tox -e benchmark`` will write on console the table (like it was done for #204 )